Add 2 weapons, add several type bullets, added Bootstrap, fixed Factory with ObjectPool, player can grab ammo and he have ammoCount for every weapon, enemy try find to player every 0.5 sec

+++ Assets/Scripts/Bullets/Models/BulletModel.cs
@@ -8,7 +8,7 @@ namespace Assets.Scripts.Bullets.Models
 {
     public abstract class BulletModel : MonoBehaviour, IObjectPool<GameObject>
     {
-        [field: SerializeField, Range(-100, 100)] public float ValueDamage { get; private set; }
+        [field: SerializeField, Range(-100, 100)] public int ValueDamage { get; private set; }
         [field: SerializeField, Range(0, 100)] public float BulletSpeed { get; private set; }
         [field: SerializeField, Range(0, 10)] public float Lifetime { get; private set; } = 2f;
         [field: SerializeField] public Vector3 ShootDirection { get; set; }
